Added an update log to keep track of changes to the game. Lots of changes, check that for information regarding v0.3

+++ Shmup/Assets/Scripts/Interfaces/IDamageable.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public interface IDamageable
 {
-    public void ReceiveDamage(int amount);
+    public void ReceiveDamage(float amount);
 
     public void Death();
 }
